refactor: 공용 컴포넌트 스타일 분리
