+++ src/components/Form/Input.tsx
@@ -4,7 +4,7 @@ import {
   RegisterOptions,
   UseFormRegister,
 } from 'react-hook-form';
-import styled from 'styled-components';
+import StyledInput from './Input.style';
 
 type InputProp<T extends FieldValues> = {
   label?: string | undefined;
@@ -39,18 +39,3 @@ function Input<T extends FieldValues>({
 }
 
 export default Input;
-
-const StyledInput = styled.div`
-  input {
-    width: 100%;
-    box-sizing: border-box;
-    padding-left: 1rem;
-    font-size: 1.6rem;
-    border: 0;
-    border-bottom: 1px solid ${({ theme }) => theme.color.gray2};
-
-    ::placeholder {
-      color: ${({ theme }) => theme.color.gray2};
-    }
-  }
-`;
